+++ .env.template
@@ -32,9 +32,6 @@ ME_CONFIG_MONGODB_URL="mongodb://${ME_CONFIG_MONGODB_ADMINUSERNAME}:${ME_CONFIG_
 ME_CONFIG_BASICAUTH_USERNAME="tnlcm"
 ME_CONFIG_BASICAUTH_PASSWORD="tnlcm"
 
-# Connect TNLCM with MongoDB
-MONGO_URI="mongodb://${MONGO_INITDB_ROOT_USERNAME}:${MONGO_INITDB_ROOT_PASSWORD}@${MONGO_HOST}:${MONGO_PORT}/${MONGO_DATABASE}?authSource=admin"
-
 # 6G-Library connection
 GIT_6GLIBRARY_HTTPS_URL="https://github.com/6G-SANDBOX/6G-Library.git"
 GIT_6GLIBRARY_BRANCH="main"
