+++ requirements.txt
@@ -4,11 +4,11 @@ flask-cors==4.0.0
 flask-restx==1.3.0
 flask-jwt-extended==4.6.0
 flask-mail==0.9.1
+mongoengine==0.28.2
 waitress==3.0.0
 python-dotenv==1.0.1
 requests==2.31.0
 python-jenkins==1.8.2
-pymongo==4.7.1
 pyyaml==6.0.1
 gitpython==3.1.43
 email-validator==2.1.1
