--- src/routes/__init__.py
@@ -1,5 +1,0 @@
-from .callback import callback_namespace
-from .sixglibrary import sixglibrary_namespace
-from .trial_network import trial_network_namespace
-from .users import users_namespace
-from .verification import verification_namespace
